REF: make pd._testing a directory
